# backend/openapi.json
{
  "components": {
    "schemas": {
      "AdmissionLineOut": {
        "properties": {
          "degree_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Degree Type"
          },
          "department": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Department"
          },
          "foreign_language": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Foreign Language"
          },
          "is_211": {
            "title": "Is 211",
            "type": "boolean"
          },
          "is_985": {
            "title": "Is 985",
            "type": "boolean"
          },
          "is_double_first_class": {
            "title": "Is Double First Class",
            "type": "boolean"
          },
          "is_self_drawn": {
            "title": "Is Self Drawn",
            "type": "boolean"
          },
          "political": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Political"
          },
          "program_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Program Code"
          },
          "program_name": {
            "title": "Program Name",
            "type": "string"
          },
          "province": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Province"
          },
          "school_name": {
            "title": "School Name",
            "type": "string"
          },
          "source_file": {
            "title": "Source File",
            "type": "string"
          },
          "subject_one": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subject One"
          },
          "subject_two": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subject Two"
          },
          "total_delta_raw": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Delta Raw"
          },
          "total_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Score"
          },
          "website": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Website"
          },
          "year": {
            "title": "Year",
            "type": "integer"
          },
          "zone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Zone"
          }
        },
        "required": [
          "year",
          "school_name",
          "program_name",
          "program_code",
          "department",
          "degree_type",
          "total_score",
          "political",
          "foreign_language",
          "subject_one",
          "subject_two",
          "total_delta_raw",
          "province",
          "is_985",
          "is_211",
          "is_double_first_class",
          "zone",
          "is_self_drawn",
          "website",
          "source_file"
        ],
        "title": "AdmissionLineOut",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ProgramSearchOut": {
        "properties": {
          "query": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Query"
          },
          "rows": {
            "items": {
              "$ref": "#/components/schemas/ProgramOptionOut"
            },
            "title": "Rows",
            "type": "array"
          },
          "truncated": {
            "title": "Truncated",
            "type": "boolean"
          }
        },
        "required": [
          "query",
          "truncated",
          "rows"
        ],
        "title": "ProgramSearchOut",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "考研择校 + 考公路径规划",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/kaoyan/admission-lines": {
      "get": {
        "operationId": "admission_lines_api_kaoyan_admission_lines_get",
        "parameters": [
          {
            "in": "query",
            "name": "school",
            "required": true,
            "schema": {
              "title": "School",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "program",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Program"
            }
          },
          {
            "in": "query",
            "name": "degree_type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Degree Type"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AdmissionLineOut"
                  },
                  "title": "Response Admission Lines Api Kaoyan Admission Lines Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Admission Lines",
        "tags": [
          "kaoyan"
        ]
      }
    },
    "/api/kaoyan/ai-report": {
      "post": {
        "operationId": "ai_report_api_kaoyan_ai_report_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AiReportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AiReportOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ai Report",
        "tags": [
          "kaoyan"
        ]
      }
    },
    "/api/kaoyan/cs408/conflicts": {
      "get": {
        "operationId": "cs408_conflicts_api_kaoyan_cs408_conflicts_get",
        "parameters": [
          {
            "in": "query",
            "name": "school",
            "required": true,
            "schema": {
              "title": "School",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Cs408ConflictOut"
                  },
                  "title": "Response Cs408 Conflicts Api Kaoyan Cs408 Conflicts Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cs408 Conflicts",
        "tags": [
          "kaoyan"
        ]
      }
    },
    "/api/kaoyan/cs408/subject-changes": {
      "get": {
        "operationId": "cs408_subject_changes_api_kaoyan_cs408_subject_changes_get",
        "parameters": [
          {
            "in": "query",
            "name": "school",
            "required": true,
            "schema": {
              "title": "School",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Cs408SubjectChangeOut"
                  },
                  "title": "Response Cs408 Subject Changes Api Kaoyan Cs408 Subject Changes Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cs408 Subject Changes",
        "tags": [
          "kaoyan"
        ]
      }
    },
    "/api/kaoyan/cs408/units": {
      "get": {
        "operationId": "cs408_units_api_kaoyan_cs408_units_get",
        "parameters": [
          {
            "in": "query",
            "name": "school",
            "required": true,
            "schema": {
              "title": "School",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "subject_class",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Subject Class"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Cs408UnitOut"
                  },
                  "title": "Response Cs408 Units Api Kaoyan Cs408 Units Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cs408 Units",
        "tags": [
          "kaoyan"
        ]
      }
    },
    "/api/kaoyan/cs408/year-lines": {
      "get": {
        "operationId": "cs408_year_lines_api_kaoyan_cs408_year_lines_get",
        "parameters": [
          {
            "in": "query",
            "name": "school",
            "required": true,
            "schema": {
              "title": "School",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "year",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Year"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Cs408YearLineOut"
                  },
                  "title": "Response Cs408 Year Lines Api Kaoyan Cs408 Year Lines Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cs408 Year Lines",
        "tags": [
          "kaoyan"
        ]
      }
    },
    "/api/kaoyan/enrollment-plans": {
      "get": {
        "operationId": "enrollment_plans_api_kaoyan_enrollment_plans_get",
        "parameters": [
          {
            "in": "query",
            "name": "school",
            "required": true,
            "schema": {
              "title": "School",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "program",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Program"
            }
          },
          {
            "in": "query",
            "name": "study_mode",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Study Mode"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/EnrollmentPlanOut"
                  },
                  "title": "Response Enrollment Plans Api Kaoyan Enrollment Plans Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Enrollment Plans",
        "tags": [
          "kaoyan"
        ]
      }
    },
    "/api/kaoyan/heat-ranking": {
      "get": {
        "operationId": "heat_ranking_api_kaoyan_heat_ranking_get",
        "parameters": [
          {
            "in": "query",
            "name": "sort",
            "required": false,
            "schema": {
              "default": "net",
              "pattern": "^(net|comp)$",
              "title": "Sort",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "province",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Province"
            }
          },
          {
            "in": "query",
            "name": "region",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Region"
            }
          },
          {
            "in": "query",
            "name": "subject_class",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Subject Class"
            }
          },
          {
            "in": "query",
            "name": "tier",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tier"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HeatRankingOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Heat Ranking",
        "tags": [
          "kaoyan"
        ]
      }
    },
    "/api/kaoyan/national-lines": {
      "get": {
        "operationId": "national_lines_api_kaoyan_national_lines_get",
        "parameters": [
          {
            "in": "query",
            "name": "year",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Year"
            }
          },
          {
            "in": "query",
            "name": "subject",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Subject"
            }
          },
          {
            "in": "query",
            "name": "kind",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Kind"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/NationalLineOut"
                  },
                  "title": "Response National Lines Api Kaoyan National Lines Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "National Lines",
        "tags": [
          "kaoyan"
        ]
      }
    },
    "/api/kaoyan/programs": {
      "get": {
        "operationId": "programs_api_kaoyan_programs_get",
        "parameters": [
          {
            "in": "query",
            "name": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Query"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProgramSearchOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Programs",
        "tags": [
          "kaoyan"
        ]
      }
    },
...
    "/api/kaoyan/rankings/program": {
      "get": {
        "operationId": "program_ranking_api_kaoyan_rankings_program_get",
        "parameters": [
          {
            "in": "query",
            "name": "year",
            "required": true,
            "schema": {
              "title": "Year",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "program_code",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Program Code"
            }
          },
          {
            "in": "query",
            "name": "program_name",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Program Name"
            }
          },
          {
            "in": "query",
            "name": "score",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Score"
            }
          },
          {
            "in": "query",
            "name": "zone",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Zone"
            }
          },
          {
            "in": "query",
            "name": "is_985",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Is 985"
            }
          },
          {
            "in": "query",
            "name": "is_211",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Is 211"
            }
          },
          {
            "in": "query",
            "name": "is_self_drawn",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Is Self Drawn"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProgramRankingOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Program Ranking",
        "tags": [
          "kaoyan"
        ]
      }
    },
...
  }
}